Fix storage structure.
